Add faux mocking
This mocking framework just came out, and it's pretty sick. I am now
able to tell a function what it will return when it is called.. Which
will allow me to test the client side of the application without setting
up a fake TCP server to give back packets for testing of the client.

Also add the Connection struct, to be mocked.

Signed-off-by: wcampbell <wcampbell1995@gmail.com>